+++ tone.py
@@ -15,7 +15,7 @@ class ToneGenerator:
         self.rate = rate
         self.harmonics_num = harmonics_num
         # create the time intervals
-        self.t = np.linspace(0, self.T, self.T*rate, endpoint=False)
+        self.t = np.linspace(0, self.T, int(self.T*rate), endpoint=False)
   
     def generate(self, f):
         '''
